+++ src/app/app.module.ts
@@ -2,20 +2,23 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppRoutingModule } from './app-routing.module';
+
 import { AppComponent } from './app.component';
 import { FirstComponent } from './first/components/first';
-import { ProductComponent } from './products/components/product';
-import { ProductListComponent } from './products/components/product-list/product-list.component';
-import { CartComponent } from './cart/components/cart.component';
+import { ProductComponent, ProductListComponent } from './products/components';
+import { CartComponent } from './cart/components';
+import { CartListComponent } from './cart/components/cart-list/cart-list.component';
+
+const declarations = [
+  AppComponent,
+  FirstComponent,
+  ProductComponent,
+  ProductListComponent,
+  CartComponent,
+];
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    FirstComponent,
-    ProductComponent,
-    ProductListComponent,
-    CartComponent
-  ],
+  declarations: [...declarations, CartListComponent],
   imports: [
     BrowserModule,
     AppRoutingModule