+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from '@angular/core';
+
+import { Product } from '../../../products/models/ProductModel.model';
+
+@Component({
+  selector: 'app-cart-list',
+  templateUrl: './cart-list.component.html',
+  styleUrls: ['./cart-list.component.scss']
+})
+export class CartListComponent {
+  @Input() cartItems: Product[];
+}