+++ src/app/products/components/index.ts
@@ -1,0 +1,2 @@
+export { ProductComponent } from './product';
+export { ProductListComponent } from './product-list';