+++ changelog.md
@@ -5,7 +5,8 @@
 3. Created ProductService that provides products array.
 4. Created ProductListComponent that shows list of products received by ProductService.
 5. Created CartComponent and used it in AppComponent.
-6. Created CartService that handles cart content:
+6. Created CartListComponent and used it in CartComponent.
+7. Created CartService that handles cart content:
    * getCartItems
    * addCartItem
    * resetCart